+++ test/AuraRouterTest.php
@@ -14,6 +14,7 @@ use Aura\Router\Map as AuraMap;
 use Aura\Router\Matcher as AuraMatcher;
 use Aura\Router\Route as AuraRoute;
 use Aura\Router\RouterContainer as AuraRouterContainer;
+use Fig\Http\Message\RequestMethodInterface as RequestMethod;
 use PHPUnit_Framework_TestCase as TestCase;
 use Psr\Http\Message\ServerRequestInterface;
 use Psr\Http\Message\UriInterface;
@@ -58,7 +59,7 @@ class AuraRouterTest extends TestCase
 
     public function testAddingRouteAggregatesRoute()
     {
-        $route  = new Route('/foo', 'foo', ['GET']);
+        $route  = new Route('/foo', 'foo', [RequestMethod::METHOD_GET]);
         $router = $this->getRouter();
         $router->addRoute($route);
         $this->assertAttributeContains($route, 'routesToInject', $router);
@@ -71,7 +72,7 @@ class AuraRouterTest extends TestCase
      */
     public function testMatchingInjectsRouteIntoAuraRouter()
     {
-        $route  = new Route('/foo', 'foo', ['GET']);
+        $route  = new Route('/foo', 'foo', [RequestMethod::METHOD_GET]);
         $router = $this->getRouter();
         $router->addRoute($route);
 
@@ -88,7 +89,7 @@ class AuraRouterTest extends TestCase
         $request->getUri()->will(function () use ($uri) {
             return $uri->reveal();
         });
-        $request->getMethod()->willReturn('GET');
+        $request->getMethod()->willReturn(RequestMethod::METHOD_GET);
         $request->getServerParams()->willReturn([]);
 
         $this->auraMap->addRoute($auraRoute)->shouldBeCalled();
@@ -103,7 +104,7 @@ class AuraRouterTest extends TestCase
      */
     public function testUriGenerationInjectsRouteIntoAuraRouter()
     {
-        $route  = new Route('/foo', 'foo', ['GET']);
+        $route  = new Route('/foo', 'foo', [RequestMethod::METHOD_GET]);
         $router = $this->getRouter();
         $router->addRoute($route);
 
@@ -121,7 +122,7 @@ class AuraRouterTest extends TestCase
 
     public function testCanSpecifyAuraRouteTokensViaRouteOptions()
     {
-        $route = new Route('/foo', 'foo', ['GET']);
+        $route = new Route('/foo', 'foo', [RequestMethod::METHOD_GET]);
         $route->setOptions(['tokens' => ['foo' => 'bar']]);
 
         $auraRoute = new AuraRoute();
@@ -142,7 +143,7 @@ class AuraRouterTest extends TestCase
 
     public function testCanSpecifyAuraRouteValuesViaRouteOptions()
     {
-        $route = new Route('/foo', 'foo', ['GET']);
+        $route = new Route('/foo', 'foo', [RequestMethod::METHOD_GET]);
         $route->setOptions(['values' => ['foo' => 'bar']]);
 
         $auraRoute = new AuraRoute();
@@ -163,7 +164,7 @@ class AuraRouterTest extends TestCase
 
     public function testCanSpecifyAuraRouteWildcardViaRouteOptions()
     {
-        $route = new Route('/foo', 'foo', ['GET']);
+        $route = new Route('/foo', 'foo', [RequestMethod::METHOD_GET]);
         $route->setOptions(['wildcard' => 'card']);
 
         $auraRoute = new AuraRoute();
@@ -190,14 +191,14 @@ class AuraRouterTest extends TestCase
 
         $request = $this->prophesize(ServerRequestInterface::class);
         $request->getUri()->willReturn($uri);
-        $request->getMethod()->willReturn('GET');
+        $request->getMethod()->willReturn(RequestMethod::METHOD_GET);
         $request->getServerParams()->willReturn([]);
 
         $auraRoute = new AuraRoute();
         $auraRoute->name('/foo');
         $auraRoute->path('/foo');
         $auraRoute->handler('foo');
-        $auraRoute->allows(['GET']);
+        $auraRoute->allows([RequestMethod::METHOD_GET]);
         $auraRoute->attributes([
             'action' => 'foo',
             'bar'    => 'baz',
@@ -206,6 +207,7 @@ class AuraRouterTest extends TestCase
         $this->auraMatcher->match($request)->willReturn($auraRoute);
 
         $router = $this->getRouter();
+        $router->addRoute(new Route('/foo', 'foo', [RequestMethod::METHOD_GET], '/foo'));
         $result = $router->match($request->reveal());
         $this->assertInstanceOf(RouteResult::class, $result);
         $this->assertTrue($result->isSuccess());
@@ -219,20 +221,20 @@ class AuraRouterTest extends TestCase
 
     public function testMatchFailureDueToHttpMethodReturnsRouteResultWithAllowedMethods()
     {
-        $route = new Route('/foo', 'foo', ['POST']);
+        $route = new Route('/foo', 'foo', [RequestMethod::METHOD_POST]);
 
         $uri = $this->prophesize(UriInterface::class);
         $uri->getPath()->willReturn('/foo');
 
         $request = $this->prophesize(ServerRequestInterface::class);
         $request->getUri()->willReturn($uri);
-        $request->getMethod()->willReturn('GET');
+        $request->getMethod()->willReturn(RequestMethod::METHOD_GET);
         $request->getServerParams()->willReturn([]);
 
         $this->auraMatcher->match($request)->willReturn(false);
 
         $auraRoute = new AuraRoute();
-        $auraRoute->allows(['POST']);
+        $auraRoute->allows([RequestMethod::METHOD_POST]);
 
         $this->auraMatcher->getFailedRoute()->willReturn($auraRoute);
 
@@ -240,19 +242,19 @@ class AuraRouterTest extends TestCase
         $result = $router->match($request->reveal());
         $this->assertInstanceOf(RouteResult::class, $result);
         $this->assertTrue($result->isFailure());
-        $this->assertSame(['POST'], $result->getAllowedMethods());
+        $this->assertSame([RequestMethod::METHOD_POST], $result->getAllowedMethods());
     }
 
     public function testMatchFailureNotDueToHttpMethodReturnsGenericRouteFailureResult()
     {
-        $route = new Route('/foo', 'foo', ['GET']);
+        $route = new Route('/foo', 'foo', [RequestMethod::METHOD_GET]);
 
         $uri = $this->prophesize(UriInterface::class);
         $uri->getPath()->willReturn('/bar');
 
         $request = $this->prophesize(ServerRequestInterface::class);
         $request->getUri()->willReturn($uri);
-        $request->getMethod()->willReturn('PUT');
+        $request->getMethod()->willReturn(RequestMethod::METHOD_PUT);
         $request->getServerParams()->willReturn([]);
 
         $this->auraMatcher->match($request)->willReturn(false);
@@ -275,9 +277,9 @@ class AuraRouterTest extends TestCase
     public function testCanGenerateUriFromRoutes()
     {
         $router = new AuraRouter();
-        $route1 = new Route('/foo', 'foo', ['POST'], 'foo-create');
-        $route2 = new Route('/foo', 'foo', ['GET'], 'foo-list');
-        $route3 = new Route('/foo/{id}', 'foo', ['GET'], 'foo');
+        $route1 = new Route('/foo', 'foo', [RequestMethod::METHOD_POST], 'foo-create');
+        $route2 = new Route('/foo', 'foo', [RequestMethod::METHOD_GET], 'foo-list');
+        $route3 = new Route('/foo/{id}', 'foo', [RequestMethod::METHOD_GET], 'foo');
         $route4 = new Route('/bar/{baz}', 'bar', Route::HTTP_METHOD_ANY, 'bar');
 
         $router->addRoute($route1);
@@ -296,14 +298,14 @@ class AuraRouterTest extends TestCase
      */
     public function testReturns404ResultIfAuraReturnsNullForFailedRoute()
     {
-        $route = new Route('/foo', 'foo', ['GET']);
+        $route = new Route('/foo', 'foo', [RequestMethod::METHOD_GET]);
 
         $uri = $this->prophesize(UriInterface::class);
         $uri->getPath()->willReturn('/bar');
 
         $request = $this->prophesize(ServerRequestInterface::class);
         $request->getUri()->willReturn($uri);
-        $request->getMethod()->willReturn('PUT');
+        $request->getMethod()->willReturn(RequestMethod::METHOD_PUT);
         $request->getServerParams()->willReturn([]);
 
         $this->auraMatcher->match($request)->willReturn(false);
@@ -323,7 +325,7 @@ class AuraRouterTest extends TestCase
     public function testGeneratedUriIsNotEncoded()
     {
         $router = new AuraRouter();
-        $route  = new Route('/foo/{id}', 'foo', ['GET'], 'foo');
+        $route  = new Route('/foo/{id}', 'foo', [RequestMethod::METHOD_GET], 'foo');
 
         $router->addRoute($route);
 
@@ -331,5 +333,63 @@ class AuraRouterTest extends TestCase
             '/foo/bar is not encoded',
             $router->generateUri('foo', ['id' => 'bar is not encoded'])
         );
+    }
+
+    public function testSuccessfulRouteResultComposesMatchedRoute()
+    {
+        $route = new Route('/foo', 'foo', [RequestMethod::METHOD_GET]);
+
+        $uri = $this->prophesize(UriInterface::class);
+        $uri->getPath()->willReturn('/foo');
+
+        $request = $this->prophesize(ServerRequestInterface::class);
+        $request->getUri()->willReturn($uri);
+        $request->getMethod()->willReturn(RequestMethod::METHOD_GET);
+        $request->getServerParams()->willReturn([]);
+
+        $router = new AuraRouter();
+        $router->addRoute($route);
+
+        $result = $router->match($request->reveal());
+        $this->assertInstanceOf(RouteResult::class, $result);
+        $this->assertTrue($result->isSuccess());
+
+        $matched = $result->getMatchedRoute();
+        $this->assertSame($route, $matched);
+    }
+
+    public function implicitMethods()
+    {
+        return [
+            'head'    => [RequestMethod::METHOD_HEAD],
+            'options' => [RequestMethod::METHOD_OPTIONS],
+        ];
+    }
+
+    /**
+     * @dataProvider implicitMethods
+     */
+    public function testHeadAndOptionsAlwaysResultInRoutingSuccessIfPathMatches($method)
+    {
+        $uri = $this->prophesize(UriInterface::class);
+        $uri->getPath()->willReturn('/foo');
+
+        $request = $this->prophesize(ServerRequestInterface::class);
+        $request->getUri()->willReturn($uri);
+        $request->getMethod()->willReturn($method);
+        $request->getServerParams()->willReturn([]);
+
+        // Not mocking the router container or Aura\Route; this particular test
+        // is testing how the parts integrate.
+        $router = new AuraRouter();
+        $route  = new Route('/foo', 'foo', [RequestMethod::METHOD_POST]);
+        $router->addRoute($route);
+
+        $result = $router->match($request->reveal());
+        $this->assertInstanceOf(RouteResult::class, $result);
+        $this->assertTrue($result->isSuccess());
+
+        $matched = $result->getMatchedRoute();
+        $this->assertSame($route, $matched);
     }
 }